# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Lumora API
  version: 0.1.0
  description: Public HTTP boundary of the Open Lumora modular monolith.
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/bundles/export:
    post:
      summary: Export selected profiles as a sanitized .lumora bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_ids]
              properties:
                agent_ids:
                  type: array
                  minItems: 1
                  items: { type: string, pattern: "^[a-z][a-z0-9]{5}$" }
                include_conversations: { type: boolean, default: false }
      responses:
        "200":
          description: Streamed, checksummed portable bundle
          content:
            application/vnd.open-lumora.bundle:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/bundles/inspect:
    post:
      summary: Validate and inspect a .lumora bundle without changing local state
      requestBody: { $ref: "#/components/requestBodies/Bundle" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/bundles/dry-run:
    post:
      summary: Preview ID remaps, resets, paused crons, quarantine, and storage use
      requestBody: { $ref: "#/components/requestBodies/Bundle" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/bundles/apply:
    post:
      summary: Atomically import a validated bundle with safe defaults
      description: Conflicting agent IDs are remapped, provider and approval state is reset, cron jobs are paused, and workspace code is quarantined.
      requestBody: { $ref: "#/components/requestBodies/Bundle" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/device/pair:
    post:
      summary: Enable registration through the configured outbound cloud endpoint
      responses:
        "202": { $ref: "#/components/responses/Success" }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/teams/:
    get:
      summary: List saved agent teams
      responses:
        "200": { $ref: "#/components/responses/Success" }
    post:
      summary: Create an owner-scoped team with bounded delegation policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamInput" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/teams/{team_id}:
    get:
      summary: Get a saved team
      parameters: [{ $ref: "#/components/parameters/TeamID" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      summary: Replace a saved team's members and execution policy
      parameters: [{ $ref: "#/components/parameters/TeamID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamInput" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
    delete:
      summary: Delete a saved team definition
      parameters: [{ $ref: "#/components/parameters/TeamID" }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /api/v1/teams/{team_id}/run:
    post:
      summary: Run bounded child work and return the orchestrator's final summary
      parameters: [{ $ref: "#/components/parameters/TeamID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task: { type: string }
                depth: { type: integer, minimum: 1, default: 1 }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  requestBodies:
    Bundle:
      required: true
      content:
        application/vnd.open-lumora.bundle:
          schema: { type: string, format: binary }
        multipart/form-data:
          schema:
            type: object
            required: [file]
            properties:
              file: { type: string, format: binary }
  parameters:
...
    TeamID:
      name: team_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Success:
      description: Successful envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    RateLimited:
      description: The active edition quota was reached
      headers:
        RateLimit-Limit:
          schema: { type: integer }
        RateLimit-Remaining:
          schema: { type: integer }
        RateLimit-Policy:
          schema: { type: string }
        RateLimit-Reset:
          schema: { type: integer, format: int64 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
  schemas:
    TeamInput:
      type: object
      required: [name, orchestrator_id, members]
      properties:
        name: { type: string }
        orchestrator_id: { type: string }
        members:
          type: array
          items:
            type: object
            required: [agent_id, role]
            properties:
              agent_id: { type: string }
              role: { type: string }
              allowed_tools:
                type: array
                items: { type: string }
              enabled: { type: boolean, default: true }
        shared_workspace: { type: boolean, default: false }
        max_parallel: { type: integer, minimum: 1 }
        max_depth: { type: integer, minimum: 1 }
        enabled: { type: boolean, default: true }
    Envelope:
      type: object
      required: [success, status_code]
      properties:
        success: { type: boolean }
        data: {}
        message: { type: string }
        status_code: { type: integer }